fix: simplify sqllineage.__init__
